+++ src/Controller/UserAccountReject.php
@@ -1,6 +1,5 @@
 <?php
-
-namespace MyModule\Controller;
+namespace MyModule\controller;
 
 use MyApp\TemplateLoader;
 use MyApp\HTTP\HTTPRequest;
@@ -11,7 +10,7 @@ use MyModule\fonction\SendEmail;
 /**
 *Class allowing deletion by denying validation of a validated user account.
 */
-class userAccountReject
+class UserAccountReject
 {
     private $templateLoader;
 
@@ -22,21 +21,31 @@ class userAccountReject
 
     public function __invoke(HTTPRequest $request)
     {
-        $data = new UserManagement;
-	    $emailAndPhoto = $data->reqUserEmailAndPhotoAccount($request->getParams()[0]);
+        if ($request->getSession('token') == $request->getGET('token')) {
+            $data = new UserManagement;
+            $emailAndPhoto = $data->reqUserEmailAndPhotoAccount($request->getGET('id'));
 
-        $content = $this->templateLoader->generate('userAccountReject.php', []);
+            $content = $this->templateLoader->generate('userAccountReject.php', []);
 
-        new SendEmail($emailAndPhoto->getEmail(), 'Désoler, refus de validation votre compte utilisateur', $content);
-	    $data->reqUserAccountReject($request->getParams()[0]);
-        //delete the photo in the avatar folder
-	    unlink(__DIR__.'./../../public/img/avatar/'.$emailAndPhoto->getPhoto());
-        $return = (new userManagement)->userAccountNoValidate();
+            new SendEmail($emailAndPhoto->getEmail(), 'Désoler, refus de validation votre compte utilisateur', $content);
+            $data->reqUserAccountReject($request->getGET('id'));
+            //delete the photo in the avatar folder
+            unlink(__DIR__.'./../../public/img/avatar/'.$emailAndPhoto->getPhoto());
+            $return = (new userManagement)->userAccountNoValidate();
 
-        $request->addSession('message', 'Le compte à été supprimé. Il recevera un message lui disant que sont compte à été refusé. Vous pouvez continué à juger des comptes utilisateurs si vous le voulez');
+            $message = 'Le compte à été supprimé. Il recevera un message lui disant que sont compte à été refusé. Vous pouvez continué à juger des comptes utilisateurs si vous le voulez';
 
-        echo $this->templateLoader->generate('userAccountManagement.php', [
-            'user' => $return,
-            'request' => $request]);
+            $this->templateLoader->twigTemplate('userAccountManagement.php', [
+                'user' => $return,
+                'request' => $request,
+                'message' => $message
+                ]);
+        } else {
+            $message = 'désolé! mais votre requête n\a pas aboutie, veuillez réessayer ultérieurement ou envoyer un email à un administrateur.';
+
+            $this->templateLoader->twigTemplate('message', [
+                'message' => $message
+                ]);
+        }
     }
 }
